+++ libs/poses/include/mrpt/poses/CPoses3DSequence.h
@@ -71,13 +71,13 @@ class CPoses3DSequence : public mrpt::serialization::CSerializable
 	 * \exception std::exception On invalid index value
 	 * \sa computeTraveledDistanceAfterAll
 	 */
-	float computeTraveledDistanceAfter(unsigned int n);
+	double computeTraveledDistanceAfter(size_t n);
 
 	/** Returns the traveled distance after ALL movements.
 	 *   A shortcut for "computeTraveledDistanceAfter( posesCount() )".
 	 * \sa computeTraveledDistanceAfter
 	 */
-	float computeTraveledDistanceAfterAll();
+	double computeTraveledDistanceAfterAll();
 
    private:
 	/** The internal sequence of poses, stored as relative, incremental poses,
